+++ R/0nly-internal.R
@@ -280,6 +280,13 @@ model.frame = function(formula, data, ...) {
     nms
 }
 
+# utility to make all names in a summary syntactically valid
+.validate.names = function(object) {
+    for (a in c("names", "pri.vars", "by.vars"))
+        attr(object, a) = make.names(attr(object, a))
+    object
+}
+
 # reorder columns of data frame to match numeric positions in num (if any)
 .reorder.cols = function(data, num) {
     if (all(is.na(num)))